# go_sprint/api/openapi/combined-with-auth.yaml
openapi: 3.0.3
info:
  title: Sprint Management API with Authentication
  version: 1.0.0
  description: Complete API for sprint management including authentication
servers:
- url: http://localhost:8082
  description: Local development server
paths:
  /api/v1/workitems:
    get:
      summary: List work items
      description: Retrieve a paginated list of work items with optional filtering
      operationId: listWorkItems
      tags:
      - Work Items
      parameters:
      - name: type
        in: query
        description: Filter by work item type
        schema:
          $ref: '#/components/schemas/WorkItemType'
      - name: status
        in: query
        description: Filter by work item status
        schema:
          $ref: '#/components/schemas/WorkItemStatus'
      - name: assignee_id
        in: query
        description: Filter by assignee user ID
        schema:
          type: string
          format: uuid
      - name: sprint_id
        in: query
        description: Filter by sprint ID
        schema:
          type: string
          format: uuid
      - name: priority
        in: query
        description: Filter by priority level
        schema:
          $ref: '#/components/schemas/PriorityLevel'
      - name: search
        in: query
        description: Full-text search across title and description
        schema:
          type: string
      - name: cursor
        in: query
        description: Cursor for pagination
        schema:
          type: string
      - name: limit
        in: query
        description: Number of items per page (max 100)
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create work item
      description: Create a new work item (epic, story, or defect)
      operationId: createWorkItem
      tags:
      - Work Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkItemRequest'
      responses:
        '201':
          description: Work item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workitems/{id}:
    get:
      summary: Get work item
      description: Retrieve a specific work item by ID
      operationId: getWorkItem
      tags:
      - Work Items
      parameters:
      - name: id
        in: path
        required: true
        description: Work item ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update work item
      description: Update an existing work item
      operationId: updateWorkItem
      tags:
      - Work Items
      parameters:
      - name: id
        in: path
        required: true
        description: Work item ID
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkItemRequest'
      responses:
        '200':
          description: Work item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete work item
      description: Soft delete a work item
      operationId: deleteWorkItem
      tags:
      - Work Items
      parameters:
      - name: id
        in: path
        required: true
        description: Work item ID
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Work item deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workitems/{id}/dependencies:
    get:
      summary: Get work item dependencies
      description: Retrieve all dependencies for a work item
      operationId: getWorkItemDependencies
      tags:
      - Work Items
      - Dependencies
      parameters:
      - name: id
        in: path
        required: true
        description: Work item ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create work item dependency
      description: Create a dependency relationship between work items
      operationId: createWorkItemDependency
      tags:
      - Work Items
      - Dependencies
      parameters:
      - name: id
        in: path
        required: true
        description: Source work item ID
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDependencyRequest'
      responses:
        '201':
          description: Dependency created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dependency'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workitems/{id}/dependencies/{dependency_id}:
    delete:
      summary: Delete work item dependency
      description: Remove a dependency relationship
      operationId: deleteWorkItemDependency
      tags:
      - Work Items
      - Dependencies
      parameters:
      - name: id
        in: path
        required: true
        description: Source work item ID
        schema:
          type: string
          format: uuid
      - name: dependency_id
        in: path
        required: true
        description: Dependency ID
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Dependency deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workitems/{id}/children:
    get:
      summary: Get child work items
      description: Retrieve all child work items (sub-tasks)
      operationId: getChildWorkItems
      tags:
      - Work Items
      - Relationships
      parameters:
      - name: id
        in: path
        required: true
        description: Parent work item ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/sprints:
    get:
      summary: List sprints
      description: Retrieve a paginated list of sprints with optional filtering
      operationId: listSprints
      tags:
      - Sprints
      parameters:
      - name: status
        in: query
        description: Filter by sprint status
        schema:
          $ref: '#/components/schemas/SprintStatus'
      - name: start_date_after
        in: query
        description: Filter sprints starting after this date
        schema:
          type: string
          format: date
      - name: start_date_before
        in: query
        description: Filter sprints starting before this date
        schema:
          type: string
          format: date
      - name: cursor
        in: query
        description: Cursor for pagination
        schema:
          type: string
      - name: limit
        in: query
        description: Number of items per page (max 100)
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create sprint
      description: Create a new sprint with date range and capacity
      operationId: createSprint
      tags:
      - Sprints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSprintRequest'
      responses:
        '201':
          description: Sprint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}:
    get:
      summary: Get sprint
      description: Retrieve a specific sprint by ID with all details
      operationId: getSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update sprint
      description: Update an existing sprint
      operationId: updateSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSprintRequest'
      responses:
        '200':
          description: Sprint updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete sprint
      description: Soft delete a sprint
      operationId: deleteSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Sprint deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/close:
    post:
      summary: Close sprint
      description: Close a sprint and move incomplete work items to backlog
      operationId: closeSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Sprint closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintCloseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/workitems:
    get:
      summary: Get sprint work items
      description: Retrieve all work items assigned to a sprint
      operationId: getSprintWorkItems
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        description: Filter by work item status
        schema:
          type: string
          enum:
          - todo
          - in_progress
          - in_review
          - done
          - blocked
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/sprints/{id}/burndown:
    get:
      summary: Get sprint burndown chart data
      description: Retrieve burndown chart data for a sprint
      operationId: getSprintBurndown
      tags:
      - Sprints
      - Metrics
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BurndownData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/velocity:
    get:
      summary: Get sprint velocity
      description: Calculate velocity for a completed sprint
      operationId: getSprintVelocity
      tags:
      - Sprints
      - Metrics
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VelocityData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/planning/capacity:
    post:
      summary: Calculate sprint capacity
      description: Calculate recommended sprint capacity based on team velocity
      operationId: calculateSprintCapacity
      tags:
      - Sprints
      - Planning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapacityCalculationRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityRecommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    WorkItemType:
      type: string
      enum:
      - epic
      - story
      - defect
      description: Type of work item
    WorkItemStatus:
      type: string
      enum:
      - todo
      - in_progress
      - in_review
      - done
      - blocked
      description: Current status of work item
    PriorityLevel:
      type: string
      enum:
      - low
      - medium
      - high
      - critical
      description: Priority level of work item
...
    WorkItemListResponse:
      type: object
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkItemSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- BearerAuth: []